requirements.txt
streamlit
groq
python-dotenv
llama-index
chromadb
pypdf

# NEW ↓
llama-index-embeddings-huggingface
sentence-transformers
transformers